backend/core-api/src/modules/AIassistant/rag-service/requirements.txt
boto3          
python-dotenv   
requests    
flask   

openai          
fastapi

pandas         
openpyxl       
numpy          
flask-cors
python-docx
faiss-cpu